Add extensions for UIImageView, and added documentation for functions
