+++ BAZProject/Extensions/Extension+UIImageView.swift
@@ -1,0 +1,23 @@
+//
+//  Extension+UIImage.swift
+//  BAZProject
+//
+//  Created by nsanchezj on 03/02/23.
+//
+
+import Foundation
+import UIKit
+
+extension UIImageView {
+    func load(url: URL) {
+        DispatchQueue.global().async { [weak self] in
+            if let data = try? Data(contentsOf: url) {
+                if let image = UIImage(data: data) {
+                    DispatchQueue.main.async {
+                        self?.image = image
+                    }
+                }
+            }
+        }
+    }
+}
