+++ BAZProject/MovieAPI.swift
@@ -12,6 +12,13 @@ final class MovieAPI {
     private let apiKey: String = "f6cd5c1a9e6c6b965fdcab0fa6ddd38a"
     let urlBase: String = "https://api.themoviedb.org/3"
 
+    
+    /**
+     obtains array movies from Api
+     - Parameter typeMovie: Is a enum for type movie
+     - Returns: An array type Movie
+     */
+    
     func getMovies(typeMovie: TypeMovie) -> [Movie]? {
         guard let url = URL(string: "\(urlBase)/movie/\(typeMovie.getOptionMovie())?api_key=\(apiKey)"),
               let data = try? Data(contentsOf: url),
@@ -21,6 +28,12 @@ final class MovieAPI {
         }
         return responseMovies.results
     }
+    
+    /**
+     obtains detail movie from Api
+     - Parameter idMovie: Is an id movie for each movie
+     - Returns: A estruct MovieDetail with more info about movie
+     */
     
     func getDetailMovie(idMovie : Int) -> MovieDetail? {
         guard let url = URL(string: "\(urlBase)/movie/\(idMovie))?api_key=\(apiKey)"),
